// src/store/reports/reports.enum.ts
enum ReportsActionTypeEnum {
  REPORTS_LOADING = "REPORTS_LOADING",
  REPORTS_LOADED = "REPORTS_LOADED",
  GET_CUSTOMER_REPORTS = "GET_CUSTOMER_REPORTS",
  GET_DRIVER_REPORTS = "GET_DRIVER_REPORTS",
  GET_VENDOR_REPORTS = "GET_VENDOR_REPORTS",
  GET_ORDERS_REPORTS = "GET_ORDERS_REPORTS",
  GET_PRODUCTS_REPORTS = "GET_PRODUCTS_REPORTS",
  GET_TRANSACTION_REPORTS = "GET_TRANSACTION_REPORTS",
  GET_INVENTORY_STOCK_REPORTS = "GET_INVENTORY_STOCK_REPORTS",
}

export default ReportsActionTypeEnum;